requirements.txt
mcp>=1.0.0
openai>=1.0.0
github3.py>=4.0.0
gitpython>=3.1.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0
jinja2>=3.1.0
markdown>=3.4.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
python-dotenv>=1.0.0 